chore: update design/functionality of copying highlighted code snippets (#7570)
* chore: design updates to copy code functionality

* add hover

* add hover sibling style

* updates to css highlight colors

* darkmode, cleanup

* update file name

* use space token

* fix css transition

* cleanup

* Fix some dark mode interaction colors

* add icon
