+++ src/styles/code.scss
@@ -37,7 +37,7 @@ code:not([class]) {
 }
 
 .pre-wrapper {
-  padding: 0 var(--amplify-space-large) 0 var(--amplify-space-xs);
+  padding: 0 var(--amplify-space-xs);
 }
 
 .pre-header {
@@ -70,13 +70,16 @@ code:not([class]) {
   -webkit-text-size-adjust: 100%;
 
   &:focus-visible {
-    outline: none;
-    box-shadow: 0 0 0 2px var(--amplify-colors-white) inset;
+    outline: 2px solid #fff;
+    outline-offset: 2px;
   }
 }
 
 .pre-code {
   flex: 1;
+  display: flex;
+  flex-direction: column;
+  --highlight-code-hover-background-color: hsl(211, 22%, 19%);
 }
 
 .token-line {
@@ -92,7 +95,15 @@ code:not([class]) {
 }
 
 .line-highlight {
+  &:first-child {
+    padding-top: var(--amplify-space-xxs);
+  }
+  &:last-child {
+    padding-bottom: var(--amplify-space-xxs);
+  }
   &:before {
+    transition: background-color
+      var(--amplify-components-button-transition-duration) ease;
     content: '';
     position: absolute;
     left: 0;
@@ -138,45 +149,101 @@ code:not([class]) {
   width: 1.8rem;
 }
 
-.highlight-copy-block {
-  all: unset;
-  width: 100%;
-  cursor: pointer;
+.highlight-copy-button {
+  background-color: var(--amplify-colors-neutral-90);
+  margin-top: var(--amplify-space-xxs);
+  border-color: transparent;
+  color: #fff;
+  align-self: flex-start;
+  border-end-end-radius: 0;
+  border-end-start-radius: 0;
+  margin-inline-start: var(--amplify-space-xxl);
   position: relative;
-}
-
-.highlight-copy-block-hint {
-  position: absolute;
-  top: 0;
-  right: 1.8rem;
-  color: white;
-}
-
-.highlight-copy-block .highlight-c4py-block-hint {
-  display: none;
-}
-
-.highlight-copy-block:focus .highlight-copy-block-hint {
-  display: block;
-}
-
-.highlight-copy-block:hover .highlight-copy-block-hint {
-  display: block;
-}
-
-.highlight-copy-block:focus .line-highlight::before {
-  background-color: var(--amplify-colors-primary-80);
-
+  user-select: none;
+  gap: var(--amplify-space-xxs);
+  &:focus {
+    box-shadow: none;
+  }
+  &:focus-visible {
+    outline: 2px solid #fff;
+    outline-offset: -2px;
+    &:after {
+      content: '';
+      position: absolute;
+      height: 2px;
+      left: 1px;
+      right: 1px;
+      bottom: -1px;
+      background-color: var(--amplify-colors-neutral-90);
+      z-index: 2;
+    }
+    & + .highlight-code {
+      outline: 2px solid #fff;
+      .line-highlight:after {
+        background-color: #fff;
+      }
+    }
+  }
+  &:hover {
+    background-color: var(--highlight-code-hover-background-color);
+    &:after {
+      background-color: var(--highlight-code-hover-background-color);
+    }
+    & + .highlight-code {
+      .line-highlight:before {
+        background-color: var(--highlight-code-hover-background-color);
+      }
+    }
+  }
   @include darkMode {
-    background-color: var(--amplify-colors-neutral-40);
+    background-color: var(--amplify-colors-neutral-20);
+    &:focus-visible {
+      &:after {
+        background-color: var(--amplify-colors-neutral-20);
+      }
+    }
+    &:hover {
+      background-color: var(--highlight-code-hover-background-color);
+      &:after {
+        background-color: var(--highlight-code-hover-background-color);
+      }
+      & + .highlight-code {
+        .line-highlight:before {
+          background-color: var(--highlight-code-hover-background-color);
+        }
+      }
+    }
   }
 }
 
-.highlight-copy-block:hover .line-highlight::before {
-  background-color: var(--amplify-colors-primary-90);
+/* This :has selector allows us to style the .highlight-copy-button
+that is preceding the .highlight-code block so that interaction between
+them looks "connected", i.e. the hover style on one triggers the 
+hover style on the other */
+.highlight-copy-button:has(+ .highlight-code:hover) {
+  background-color: var(--highlight-code-hover-background-color);
+  &:focus-visible {
+    &:after {
+      height: 4px;
+      bottom: -3px;
+      background-color: var(--highlight-code-hover-background-color);
+    }
+  }
+  &:hover {
+    &:after {
+      background-color: var(--highlight-code-hover-background-color);
+    }
+  }
+}
 
-  @include darkMode {
-    background-color: var(--amplify-colors-primary-10);
+.highlight-code {
+  position: relative;
+  margin-bottom: var(--amplify-space-xxs);
+  &:hover {
+    cursor: pointer;
+    .line-highlight:before {
+      background-color: var(--highlight-code-hover-background-color);
+    }
   }
 }
 
@@ -188,6 +255,7 @@ code:not([class]) {
   padding-block: var(--amplify-space-xxxs);
   color: var(--amplify-colors-white);
   border-radius: var(--amplify-radii-small);
+  gap: var(--amplify-space-xxs);
   &:hover {
     color: var(--amplify-colors-white);
     background-color: var(--code-copy-hover-background-color);
